Added the ability to specify templates in view pages rather than in Python view classes.
Added the ability to supply a class when defiening views from
templates. In this case, the class becomes a base class of the
generated view class.

Added the ability to supply base classes to simple view classes.

Changed the zope page-template engine to use
Zope.App.Traversers.Traverser.Traverser directly, wo looking it up as
an adapter. I suspect that we'll stop making this pluggable, as it's
too much bother, given that adapters are used for each step anyway.


git-svn-id: svn://svn.zope.org/repos/main@5959 62d5b8a3-27da-0310-9561-8e5933582275