bucket_split():  Don't leak the keys if memory for values can't be found.
PyMalloc(), PyRealloc():  Call malloc() and realloc() directly instead
of PyMem_Malloc() and PyMem_Realloc().  If the latter are used, then
memory must be freed via PyMem_Free(), but PyMem_Free() isn't called
anywhere -- memory is released via raw system free().  It would probably
be better to change all uses of free() instead, but that's A Project (well,
compared to this ...).


git-svn-id: svn://svn.zope.org/repos/main@1921 62d5b8a3-27da-0310-9561-8e5933582275