Be more cautious open creating a file when the old one doesn't exist.
Workaround for bug in Python versions upto and including 2.1.3.  If
the interpreter was builtin without large file support but the
platform had it, os.path.exists() would return false for large files.
The fix is to try to open the file first, and decide whether to create
a new file based on errno.

(Only the two of the last three chunks of the diff affect this
behavior.)

Also cleanup imports and add some whitespace between arguments to func
calls.


git-svn-id: svn://svn.zope.org/repos/main@1937 62d5b8a3-27da-0310-9561-8e5933582275