RDB/SQL Sprint team (Vika, Michael and Stephan):
Alright, to make nice dynamic SQL methods we really want DTML ported for many reasons. This is the first cut at the problem. The DTML code is based on the Zope 2.3 release version. We have implemented some initial tests for the directives and ripped out all of the old ZMI and security cruft.


git-svn-id: svn://svn.zope.org/repos/main@6074 62d5b8a3-27da-0310-9561-8e5933582275