Make tracebacks clickable.
A traceback shown for a given error or window now has each stack entry
clickable; clicking on it will invoke your editor.

The command to launch your editor will be taken from one of the environment
variables PYUNIT_EDITOR, EDITOR_REMOTE or EDITOR (in that order), and it
will be passed the filename and the line number of the stack item. The
command line constructed is:

  command +line filename

This command line works well with both Vi and Emacs; if EDITOR_REMOTE is set
to 'gvim --remote-silent' for example, traceback stack items will open up in
an existing Vim window on the correct line when clicked.

For editors that use different commandline options for specifying the line
to open can use a small shell script to translate arguments.


git-svn-id: svn://svn.zope.org/repos/main@6081 62d5b8a3-27da-0310-9561-8e5933582275