Fixed CMFDefault.RegistrationTool to correctly check the lack of 'email' property when creating a new member or when checking member's properties validity (Tracker #508).
The fix ensures that the 'email' property is filled in at Join time, and
that a user is not allowed to clear his email (the policy in CMFDefault
seems to be that all members have an email).

I keep the possibility of a user leaving an already-empty password blank
when changing his preferences (for instance if it was removed by the
administrator for some reason, the user can still go change his other
properties).


git-svn-id: svn://svn.zope.org/repos/main@34979 62d5b8a3-27da-0310-9561-8e5933582275