Correctly reindex the just-reset workflow variables of an object in a DCWorkflow after a paste.
(I put the responsibility for this in CMFCatalogAware.manage_afterClone
and not in notifyCreated because the latter is called from other places
that already have do the reindexing for other reasons.)

Cleaned up DefaultWorkflow so that the notifySuccess & co methods are
not called twice, the WorkflowTool deals with that. Reindexing after a
workflow transition is also taken care of by WorkflowTool.


git-svn-id: svn://svn.zope.org/repos/main@34990 62d5b8a3-27da-0310-9561-8e5933582275