Reimplemented service managers to be package based. Service managers are no longer containers. They have a packages subobject (not a packages service) that contains packages. TTW components are created in packages. To register a component, create the appropriate component directive objects (these should be called configuration objects).
This should be viewed as a prototype to illustrate the idea of
packages. Lots of things can change (especially UI) and many things
aren't done (e.g. visiting created directives).

In the course of this, I fixed a bunch of bugs and problems in
traversal machinery.

I also renamed Zope.ComponentArchitecture.IServiceManager back to
IServiceService, since this interface doesn't actually specify any
management.


git-svn-id: svn://svn.zope.org/repos/main@6112 62d5b8a3-27da-0310-9561-8e5933582275