During the Zope3 sprint in Charleroi, it was decided that the contract of the validate() method was not to ever convert. It would raise an exception if validation fails, and otherwise pass silently. We refactored the code accordingly. This was inadvertently changed by Stephan back to the original behavior (to assume conversion) in his recent checkins when he "finished the Schema code".
Unfortunately I do not think we stated this explicitly in the
interface, though checking CVS entry messages or talking to the people
who last worked on the schema would've spared Stephan and now me a lot
of useless work. :(

Whitespace normalization, the only validator which actually seemed to
be doing any conversion (changing its input), belongs in the converter
code, not in the validator code. Where the converter code belongs is
another debate. My opinion is that while which conversions to use is
part of the particular view, many common conversions also make sense to
be part of the schema package.


git-svn-id: svn://svn.zope.org/repos/main@6143 62d5b8a3-27da-0310-9561-8e5933582275