Beginnings of a refactoring of Validator. Mostly renaming and some changes to the hierarchy. Actually I hope the hierarchy inside Validator can mostly go away in favor of one of Fields, later (maintaining two related hierarchies is a pain).
Also some bugfixes (missing ValidationError import) to the tests. Though
now I'm wondering whether we need a try.. except ValidationError in the
tests at all; it seems to obscure the tests more than they help, so removing
the 'try' and 'except' and simply letting the block of code unprotected
may be better -- any unexpected ValidationError will cause the test to
fail with an error.


git-svn-id: svn://svn.zope.org/repos/main@6146 62d5b8a3-27da-0310-9561-8e5933582275