Cleanup tests:
Remove unused unittest boilerplate.

Use full path in import rather than implicit one, which caused some
extra confusion because the tests package and the parent have a module
with the same name.

Rename the test class in PropertySetDef to be explicit about its
purpose.  This eliminates the name for an import as in the unittest.


git-svn-id: svn://svn.zope.org/repos/main@6157 62d5b8a3-27da-0310-9561-8e5933582275