Import get_transaction() from Transaction.
XXX I'm not sure if the change in ZopePublication.py is correct,
because it was merely testing for the presence of get_transaction() in
builtins.  Since that behavior is going away, I'm not sure if it
should always import get_transaction() or always use its stub
implementations.  The former makes more sense.


git-svn-id: svn://svn.zope.org/repos/main@6185 62d5b8a3-27da-0310-9561-8e5933582275