Merged change from HEAD:
Improved the ActionsTool so that permission checking for actions is done
in the context of the object for all categories that start with 'object'
or 'workflow', in the context of the folder for those that start with
'folder', and in the context of the portal otherwise.

This is useful for admin-defined additional categories.


git-svn-id: svn://svn.zope.org/repos/main@35105 62d5b8a3-27da-0310-9561-8e5933582275