Revise Transaction API based on some discussion on persistence SIG.
The core of the change is that objects register with their data
managers and data managers join the transaction.  This simplifies the
bookkeeping done by the transaction.

The new API is based on the TP monitor API discussed in Transaction
Processing by Gray and Reuter.  The rough correspondence between the
old API and the new API is:

prepare() was tpc_begin() + commit() + tpc_vote()
commit() was tpc_finish()
abort() was tpc_abort()
savepoint() was commit(1)

The abort() and commit() methods of the old API are gone.  It's up to
the data managers to do the per-object accounting.

The basic Transaction package now defines a Transaction object without
the ZODB metadata fields.  ZODB will grown its own Transaction that
extends the base class.

Use an explicit Manager class and subclass that implements policy
associating each thread with a different transaction.


git-svn-id: svn://svn.zope.org/repos/main@6206 62d5b8a3-27da-0310-9561-8e5933582275