Merged changes from HEAD:
Made TypesTool and WorkflowTool rely exclusively on _getPortalTypeName()
to determine the portal type of an object. If the method doesn't exist
or returns None, then we're not looking at a portal object.

Now meta_type is not used as a fallback anymore. Portal objects should
be properly initialized if they want to participate in the portal types
or workflows.

Also ensured that the CMF root is not workflowable by setting its
portal_type to None.


git-svn-id: svn://svn.zope.org/repos/main@35268 62d5b8a3-27da-0310-9561-8e5933582275