Work around BTrees bug.
Some revision of the the BTrees code (possibly even the current head)
has a bug which causes (in this case) an OOBTree to lie about its
items.

Code like:

for key in oobtree.keys(): oobtree[key]

... will cause a KeyError when oobtree is asked to return the object
represented by key.

We work around this by explicitly detecting this problem and logging.


git-svn-id: svn://svn.zope.org/repos/main@22648 62d5b8a3-27da-0310-9561-8e5933582275