applying the skin to the REQUEST was done in the SkinsContainer. moved this out to the Skinnable.py class (PortalObject).  where it now has a seperate method called changeSkin(name).  The skinnable object will now do the apply of the skins to the PortalObject._v_skindata
We do this so we can change the skin during a REQUEST.  the setupCurrentSkin
stopped work when shane introduced some optimization.  I'm not quite sure
how 'unoptimized' this change is.  but it works.  I posted to plone-dev mailing list
on how to apply a skin conditionally.

main concept: we want to be able to see the environment and 'toggle' the skin
through a Script(Python) bound by a Access Rule


git-svn-id: svn://svn.zope.org/repos/main@35297 62d5b8a3-27da-0310-9561-8e5933582275