Quick hack fix to these tests.  XXX I don't really understand them.
The HoserStoppage and ExceptionInTpcFinish were failing intermittently
on Windows.  I assume the problem is that a single error in tpc_finish()
can be recovered from iff the error is for the first jar.  I assume that
most of the time the error occurred after some other jar had committed.
Fixed the test by causing two different jars to fail.


git-svn-id: svn://svn.zope.org/repos/main@2001 62d5b8a3-27da-0310-9561-8e5933582275