Fix testHoserStoppage(), and remove testExceptionInTpcFinish() and BUGtestExceptionInSubTpcBegin().
The tpc_finish test removed duplicated the logic of "hoser stoppage,"
while the BUGtest was never executed.

The testHoserStoppage() test used to depend on the order of objects in
a dict's values() list.  This order was unpredictable and caused
intermittent failures.  Make two changes to work around this:  1) Mark
the non-error object as modified first, since this seems to get
values() in a good order.  2) Retry the test with 2 randomly selected
objects until it passes, but fail after 10 tries.


git-svn-id: svn://svn.zope.org/repos/main@2003 62d5b8a3-27da-0310-9561-8e5933582275