Small but important correction to condition variable protocol!
After setting self._transaction, release the lock.  Then reacquire it
before the notify().  The wait() call is sufficient to block the other
threads; holding the lock is not necessary.

As a result, it doesn't matter if the same thread calls tpc_begin()
and tpc_abort() for a given transaction.  So remove assertion.


git-svn-id: svn://svn.zope.org/repos/main@2007 62d5b8a3-27da-0310-9561-8e5933582275