Allow read-only client connections to call pack().
Not sure about this judgement call.  It is really convenient to be
able to pack a ZEO server, even if it is only accepting read-only
connections.  Is pack() a write operation?  At one level: Yes, it
modifies the data.fs.  But it doesn't modify current objects, so it
isn't the same as calling store().

We can revisit this later if it ends up being a problem for someone.


git-svn-id: svn://svn.zope.org/repos/main@2010 62d5b8a3-27da-0310-9561-8e5933582275