Another attempt to make testHoserStoppage work consistently.
The order in which the tpc_finish() method of the jars is called is
dependent on the order in which objects are returned from
dict.values().  There is no way to write a test that works on all
platforms that produces a consistent result in dict.values(), so stop
trying.

Instead, have the jars explicitly coordinate their tpc_finish()
methods so that the first call always succeeds and the second call
always fails.


git-svn-id: svn://svn.zope.org/repos/main@2017 62d5b8a3-27da-0310-9561-8e5933582275