Handle abortVersion() correctly with all versions of ZODB.
Different versions of ZODB do different things with the serialnos for
abortVersion().  To be safe and avoid giving non-version data with an
invalid serialno after an abort version, invalidate the version and
non-version data.

This hurts cache effectiveness when a version is aborted, but I expect
it will have little practical impact.


git-svn-id: svn://svn.zope.org/repos/main@2114 62d5b8a3-27da-0310-9561-8e5933582275