I don't like subsystem or label names with spaces in them.  Instead of "ClientStorage <pid>", use "ClientStorage:<pid>".
Also log a message when a ClientStorage instance is created, showing
the pid, the read_only and read_only_fallback parameters, and the
storage name.

Also log some more things: the retry of the register() call with
read_only=1, the notifyConnected call,

Also clean up a few comments and docstrings, removing one XXX that
seems out of date.

Also change the _tpc_cond.wait() method call into a wait(30).  That
way it is interruptable.


git-svn-id: svn://svn.zope.org/repos/main@2169 62d5b8a3-27da-0310-9561-8e5933582275