Import Disconnected from the correct module.
Add a LOG call to ConnectionTests.setUp().

Add tests for the ClientStorage read_only and read_only_fallback
features.  Note that the latter feature is not completely working yet;
the reconnect feature doesn't work, so it's of questionable value for
now (though it *does* connect to a read-only storage whe no read-write
storage is available).


git-svn-id: svn://svn.zope.org/repos/main@2171 62d5b8a3-27da-0310-9561-8e5933582275