Added a new view, 'SelectedManagementView', that redirects to the first view that a user can access. Changed between-object navigation urls, like those in breadcrumbs and the contents management view, to use this view.  This corresponds to Zope 2's 'manage_workspace' method.
Changes the name of the folder contents management view to
'contents.html'. Added a new 'index.html' view as the default view for
folders. If the folder contains an object named 'index.html', then the
view redirects to that object. Otherwise, the view displays a folder
listing with linds to folder items.

Changed the view order for a number of content types so that
management views come before the default view.  This means you can now
traverse to a file or page template from a contents listing and get an
editing view.


git-svn-id: svn://svn.zope.org/repos/main@6333 62d5b8a3-27da-0310-9561-8e5933582275