Made a start with more docstrings.  But I got interrupted by the following item:
Picklers are not re-entrant!  The only place where the shared global
pickler was used was in the error handling in store(), to test whether
the caught exception can be pickled (else it would fail when sending
the response).  Change the flow control to create a new pickler
locally, only when an exception is caught.  (When no exception is
caught, we know newserial is a string so there's no need to test its
picklability.)


git-svn-id: svn://svn.zope.org/repos/main@2279 62d5b8a3-27da-0310-9561-8e5933582275