Make return addresses *@zope.org, rather than *@cvs.zope.org.
Also, applied fix to be deterministic about binary files, according to
whether the file being checked in has binary-mode keyword expansion
set.

Since there is no CVS interface command that will directly query the
repository file for this bit, we read the archive file directly
looking for the right line (which must be in the first section of the
file, so it's a short and easy check).


git-svn-id: svn://svn.zope.org/repos/main@32259 62d5b8a3-27da-0310-9561-8e5933582275