Changed standard_macros to provide macros aggregated from view_macros.pt and also from dialog_macros.pt.
Introduced the notion of a "dialog", which is like a normal ZMI view,
except that management tabs (and other such links to views) are missing.
This makes all the links on the Add-form work.
You make something a dialog by using the macro standard_macros/dialog
rather than standard_macros/page.


git-svn-id: svn://svn.zope.org/repos/main@6550 62d5b8a3-27da-0310-9561-8e5933582275