Added a facility to make it easy to get to Zope 3 from a Python prompt. Also added rudimentary debugging support.
From lib/python (or with lib/python in your Python path), start Python
and:

  from Zope.App import Application

  # Create an Application object
  app = Application('../../Data.fs', '../../site.zcml')

to get the applicatiuon root object on an open database connection,
call the application object:

  root = app()

to execute a request for a path:

  app.debug(path)

You can pass various extra parameters to debug to specify environment
variables and so on. (For now, see lib/python/Zope/App/_app.py.)

To prevent exceptions from being caught, provide the pm keyword
argument:

  app.debug(path, pm=1)

This will let you do pdb post-mortem debugging of errors.

Added a little bit of framework for site initialization, so that the
debugging environment and z3.py use the same initialization code.


git-svn-id: svn://svn.zope.org/repos/main@6598 62d5b8a3-27da-0310-9561-8e5933582275