This checkin cleans up the ObjectHub system.
First of all, the Zope.ObjectHub and Zope.App.OFS.Services.LocalObjectHub packages are gone, replaced by Zope.App.OFS.Services.ObjectHub, as per discussion with Jim and Steve.

Second, the hub events have been modified to match Jim's approach with the ObjectEvents (i.e., events are usually handed an object at the get go whenever possible).  Incidentally, this also coincides with making the "moved" hub event implementation actually coincide with the interface (by including fromLocation).  This is as per discussion with Jim.

Third, lookupLocation and lookupHubid have been switched to getLocation and getHubid.  This is a bit of a ninja-checkin or whatever the term is since I didn't bandy this change about beforehand.  :-(  Sorry.  If folks say nay then I will take responsibility for removing this change.

I think that's about it.


git-svn-id: svn://svn.zope.org/repos/main@6635 62d5b8a3-27da-0310-9561-8e5933582275