A port of RAMCacheManager to Zope3.
The current architecture is very much based on the Zope2 RAMCacheManager,
so it might use some refactoring in the future.  For instance, two
different caching interfaces can be derived: one for the data (which does
not care for request) and one for the views (which treats request data in
some special way). The current implementation should work both ways.


git-svn-id: svn://svn.zope.org/repos/main@6657 62d5b8a3-27da-0310-9561-8e5933582275