Merge the rdmurray-metameta-branch.
This brings in a number of changes to the configuration system.  The
update to doc/zcml/meta.zcml documents the "new way" of doing
meta-configuration.  The old way still works fine, and the only
meta.zcml file I've changed to the new way (so far) is the one
for the content directive.

This update introduces a new utility, makezcmldocs.py.  It is best
run under python2.3, because under 2.2 it uses a fake TextWrap class
that is very stupid about wrapping.  If you run makezcmldocs.py
it will extract any existing documentation from the metaconfiguration
directives in the system and write out structured text files in a
directory named 'doc/zcml.new/namespaces.zope.org'.  (Eventually
this will be changed so that it replaces the existing
doc/zcml/namespaces.zope.org).  It also spits out error messages
for any directives that don't have description attributes.  If yours
are among them, please consider fixing them <grin>.

One other change introduced by this update that it is important
to be aware of is that INonEmptyDirectives *must* now declare
that they implement INonEmptyDirective or they will not function
correctly.  (Previous to this update they would work as long
as at least one subdirective was actually defined).  All of
the existing directives are updated in this checkin.

Also, note that if you do fix your meta.zcml files, and your unit
tests actually test your meta.zcml file, you will need to update
your unit test to call XMLConfig('metameta.zcml',Zope.Configuration)()
before reading your own meta.zcml file.  This update is done for
the content directive in this checkin.


git-svn-id: svn://svn.zope.org/repos/main@6676 62d5b8a3-27da-0310-9561-8e5933582275