This work by brad@codeworks.lt
I created two more permissions, one of which I granted to a principal, and the
other that I denied from him so as to tickle a few more paths in the code that
was being tested. This test now exercises the errors in ZopeSecurityPolicy
that Steve fixed a short while ago.


git-svn-id: svn://svn.zope.org/repos/main@6694 62d5b8a3-27da-0310-9561-8e5933582275